Replace error.StopIteration!T with ?T
StopIteration is a pythonism that makes things a twidge clunkier than
they ought to be, especially for the common case of a coroutine that
yields data till it does not (as opposed to one with an intermittent
bi-directional data stream like Python's iterators support).

This commit replaces StopIteration with optional types. This is less
pythonic, more zigish, a bit more convenient for common use cases, and a
bit less convenient for uncommon use cases.